# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Customer service chat system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/agents:
    get:
      operationId: adminListAgents
      tags: [admin]
      summary: List all agents (super_admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAgent"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
    post:
      operationId: adminCreateAgent
      tags: [admin]
      summary: Create a new agent (super_admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgent"
        "400":
          description: Validation error
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "409":
          description: Username already taken
  /admin/agents/{id}:
    patch:
      operationId: adminUpdateAgent
      tags: [admin]
      summary: Update agent profile (super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgent"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: adminDeleteAgent
      tags: [admin]
      summary: Delete an agent (super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /admin/transfers:
    get:
      operationId: adminListTransfers
      tags: [admin]
      summary: List recent session transfers (super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Transfer records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransfersResponse"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
  /admin/logs:
    get:
      operationId: adminListLogs
      tags: [admin]
      summary: List system audit logs (super_admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: action
          in: query
          schema:
            type: string
      responses:
        "200":
          description: System log records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLogsResponse"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SessionTransfer:
      type: object
      required: [id, sessionId, fromAgentId, toAgentId, initiatedBy, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        fromAgentId:
          type: integer
        toAgentId:
          type: integer
        initiatedBy:
          type: integer
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        fromAgentDisplayName:
          type: ["string", "null"]
        toAgentDisplayName:
          type: ["string", "null"]
        initiatedByDisplayName:
          type: ["string", "null"]
        visitorNickname:
          type: ["string", "null"]
...
    AdminTransfersResponse:
      type: object
      required: [transfers]
      properties:
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/SessionTransfer"
    SystemLog:
      type: object
      required: [id, action, createdAt]
      properties:
        id:
          type: integer
        actorId:
          type: ["integer", "null"]
        actorUsername:
          type: ["string", "null"]
        actorRole:
          type: ["string", "null"]
        action:
          type: string
        targetType:
          type: ["string", "null"]
        targetId:
          type: ["integer", "null"]
        detail:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AdminLogsResponse:
      type: object
      required: [logs]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/SystemLog"
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    AdminAgent:
      type: object
      required: [id, username, role, displayName, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [agent, super_admin]
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        introduction:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateAgentBody:
      type: object
      required: [username, password, displayName]
      properties:
        id:
          type: integer
          minimum: 1
          description: Optional custom agent ID. Auto-assigned when omitted.
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        displayName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [agent, super_admin]
        introduction:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    UpdateAgentBody:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [agent, super_admin]
        introduction:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        password:
          type: ["string", "null"]
          minLength: 6
...